+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Released changes are shown in the
 
 ### Added
 - New class analysis section, with class overlap detection. More details are available in the User Guide and in the Key Concepts.
+- Pipeline pre/post-processing steps breakdown in the utterance detail page.
 
 ### Changed
 - The default port for the front end will now be 3000, the React default, instead of 3005.